Compile successful

+++ src/modules/interface/reference_generator.h
@@ -5,7 +5,7 @@
 #include <stdbool.h>
 #include <stdint.h>
 
-extern float zAccDesired;
+extern uint16_t zAccDesired;
 extern float eulerRollDesired;   // Measured roll angle in deg
 extern float eulerPitchDesired;  // Measured pitch angle in deg
 extern float eulerYawDesired;    // Measured yaw angle in deg
